requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
requests>=2.28.2
beautifulsoup4>=4.12.0
lxml>=4.9.2
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.1